+++ src/components/CarouselSection/GridCarousel.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { MDBCol, MDBRow, MDBContainer, MDBMask, MDBView } from 'mdbreact';
+import PropTypes from 'prop-types'
 function GrideCarousel({ activeCardItem }) {
   let booksArray = [
     {
